Create Median of Three.java
Your task is to program an algorithm that for three distint integers, returns their median. For example, if the input is 3, 1 and 2, the output is 2. You only need to implement the method "Median" in the file "Main.java".

Great thanks to Striker Badger, who explained the functions and conditions of the code.